requirements.txt
streamlit>=1.24
sentence-transformers==2.2.2
faiss-cpu
numpy>=1.24
pandas>=2.0
lightgbm>=4.0
scikit-learn>=1.2
pypdf>=3.0
python-docx>=0.8.11
ssdeep==3.4
python-magic
tqdm